requirements.txt
numpy==1.26.4
streamlit==1.31.1
tensorflow==2.15.0
pillow==10.2.0
streamlit-image-select==0.6.0
twilio==9.1.0
opencv-python-headless
